Update reviews dialog styles
