+++ app/frontend/containers/ReviewsDialogContainer.js
@@ -2,11 +2,14 @@ import { connect } from 'react-redux';
 import ReviewsDialog from '../ui/ReviewsDialog';
 import closeReviewsDialog from '../actions/closeReviewsDialog';
 import openReviewDialog from '../actions/openReviewDialog';
+import selectPlaceForReview from '../actions/selectPlaceForReview';
 
 const mapStateToProps = state => {
   return {
     dialogOpen: state.mapDetail.reviewsDialogOpen,
-    reviews: state.mapDetail.spotReviews
+    reviews: state.mapDetail.spotReviews,
+    currentMap: state.mapDetail.currentMap,
+    currentSpot: state.spotCard.currentSpot,
   };
 };
 
@@ -18,6 +21,14 @@ const mapDispatchToProps = (dispatch, ownProps) => {
 
     handleReviewClick: (review) => {
       dispatch(openReviewDialog(review));
+    },
+
+    handleAddReviewButtonClick: spot => {
+      let place = {
+        description: spot.name,
+        placeId: spot.place_id
+      };
+      dispatch(selectPlaceForReview(place));
     }
   };
 };
